// src/renderer/src/components/tablelib/Model/PublicModel.ts
export interface CellTemplate<TCell extends Cell = Cell> {
  /**
   * Validates and converts `uncertainCell` to compatible cell type
   *
   * @param {Uncertain<TCell>} uncertainCell Cell with all optional fields of its base (`TCell`)
   * @returns {Compatible<TCell>} Compatible cell of its base (`TCell`)
   */
  getCompatibleCell(uncertainCell: Uncertain<TCell>): Compatible<TCell>

  /**
   * Returns `true` if the cell is focusable
   *
   * @param {Compatible<TCell>} cell Current cell as `Compatible` cell
   * @returns {boolean} `true` if cell should be focusable, by default returns `true`
   */
  isFocusable?(cell: Compatible<TCell>): boolean

  /**
   * Updates cell based on new props. If not implemented, cell will be read-only
   *
   * @param {Compatible<TCell>} cell Current cell
   * @param {UncertainCompatible<TCell>} cellToMerge Incoming cell
   * @returns {Compatible<TCell>} Merged cell as `Compatible` cell
   */
  update?(cell: Compatible<TCell>, cellToMerge: UncertainCompatible<TCell>): Compatible<TCell>

  /**
   * Handles keydown event on cell template and double click (opening cell in edit mode)
   * Default: cell => { cell, enableEditMode: false }
   *
   * @param {Compatible<TCell>} cell Incoming `Compatible` cell
   * @param {number} keyCode Represents the key pressed on the keyboard, or 1 for a pointer event (double click).
   * @param {boolean} ctrl Is `ctrl` pressed when event is called ()
   * @param {boolean} shift Is `shift` pressed when event is called
   * @param {boolean} alt Is `alt` pressed when event is called
   * @param {string} [key] Represents the value of the key pressed by the user. Optional for backwards compatibility.
   * @param {boolean} capsLock Is caps lock active when event is called. Optional for backwards compatibility.
   * @returns {{ cell: Compatible<TCell>; enableEditMode: boolean }} Cell data and edit mode either affected by the event or not
   */
  handleKeyDown?(
    cell: Compatible<TCell>,
    keyCode: number,
    ctrl: boolean,
    shift: boolean,
    alt: boolean,
    key?: string,
    capsLock?: boolean
  ): { cell: Compatible<TCell>; enableEditMode: boolean }

  /**
   * Handles compositionEnd event on cell template (opening cell in edit mode)
   *
   * @param {Compatible<TCell>} cell Incoming `Compatible` cell
   * @param {string} eventData The characters generated by the input method that raised the event
   * @returns {{ cell: Compatible<TCell>; enableEditMode: boolean }} Cell data and edit mode either affected by the event or not
   */
  handleCompositionEnd?(
    cell: Compatible<TCell>,
    eventData: string
  ): { cell: Compatible<TCell>; enableEditMode: boolean }

  /**
   * Returns custom styles based on cell data applied to the cells `div` element
   * Default: _ => cell.style | {}
   *
   * @param {Compatible<TCell>} cell Incoming `Compatible` cell
   * @param {boolean} isInEditMode Flag is set to `true`, if cell is rendered in edit mode
   * @returns {CellStyle} Custom cell styling properties
   */
  getStyle?(cell: Compatible<TCell>, isInEditMode: boolean): CellStyle

  /**
   *  Returns CSS classes based on cell data applied to the cells `div` element
   *
   * @param {Compatible<TCell>} cell Incoming `Compatible` cell
   * @param {boolean} isInEditMode Flag is set to `true`, if cell is rendered in edit mode
   * @returns {string} Cells CSS class names
   */
  getClassName?(cell: Compatible<TCell>, isInEditMode: boolean): string

  /**
   * Renders the cell content
   *
   * @param {Compatible<TCell>} cell Incoming `Compatible` cell
   * @param {boolean} isInEditMode Flag is set to `true`, if cell is rendered in edit mode
   * @param {(cell: Compatible<TCell>, commit: boolean) => void} onCellChanged Callback used for commiting changes on a cell. For example: typing on html `input` element
   * @returns {React.ReactNode} Content of a cell
   */
  render(
    cell: Compatible<TCell>,
    isInEditMode: boolean,
    onCellChanged: (cell: Compatible<TCell>, commit: boolean) => void
  ): React.ReactNode
}

/**
 * `Id` is a common type to identify many ReactGrids objects
 *
 * @see https://reactgrid.com/docs/3.1/7-api/1-types/4-id/
 */
export type Id = number | string
// ...
export interface BorderProps {
  /** Color of border - e.g. `#eee`/`red` */
  readonly color?: string
  /** Style of border - e.g. `dotted`/`solid` */
  readonly style?: string
  /** Width of border - e.g. `2px` */
  readonly width?: string
}

/**
 * This interface styles single cell and prevents passing unwanted CSS properties that could break down grid rendering
 *
 * @see https://reactgrid.com/docs/3.1/7-api/0-interfaces/7-cell-style/
 */
export interface CellStyle {
  /** CSS `color` property */
  readonly color?: string
  /** CSS `background` property */
  readonly background?: string
  /** CSS `overflow` property */
  readonly overflow?: string
  /** CSS `padding-left` property */
  readonly paddingLeft?: string
  /** Object that contains all cell's borders properties */
  readonly border?: {
    readonly left?: BorderProps
    readonly top?: BorderProps
    readonly right?: BorderProps
    readonly bottom?: BorderProps
  }
}
// ...
export interface Cell {
  /** Name of cell type, must be unique */
  type: string
  /** Marks cell as non editable (by default: `false`) */
  nonEditable?: boolean
  /** `Id` of group to which this cell belongs to */
  groupId?: Id
  /** Allowed style properties contained in `CellStyle` interface */
  style?: CellStyle
  /** Additional CSS classes */
  className?: string
}

/**
 * Cell type marker - every field of `TCell` is optional.
 * Cell of this type will have only one essential field provided by `Cell` interface - `type`.
 *
 * @see https://reactgrid.com/docs/3.1/7-api/1-types/2-uncertain-cell/
 */
export type Uncertain<TCell extends Cell> = Partial<TCell> & Cell

/**
 * Cell type marker - extended & exchangeable cell (compatible with different types)
 *
 * @see https://reactgrid.com/docs/3.1/7-api/1-types/1-compatible-cell/
 */
export type Compatible<TCell extends Cell> = TCell & {
  /** Text value of a cell */
  text: string
  /**  Numeric value of a cell, if there is no numeric value representation use `NaN` */
  value: number
}

/**
 * `UncertainCompatible` is a cell type that is compatible with other cell types
 * that can be instances of various cell types (e.g. `DataCell` and `TimeCell`).
 *
 * @see https://reactgrid.com/docs/3.1/7-api/1-types/3-uncertain-compatible-cell/
 */
export type UncertainCompatible<TCell extends Cell> = Uncertain<TCell> & {
  /** Text value of a cell */
  text: string
  /** Numeric value of a cell, if there is no numeric value representation use `NaN` */
  value: number
}